refactor: update tests, multiple minor fixes
