+++ platform/api/tests/core/test_keystore.py
