+++ platform/api/tests/core/test_mailer.py
