+++ platform/api/tests/core/test_settings.py
