+++ platform/api/tests/core/test_spa.py
