+++ platform/api/tests/domains/auth/test_auth.py
